fix intermittent segfault when pending future dropped
When fdb_future_destroy() is called on a future that hasn't resolved,
it causes the callback to fire. If it fires after the box the waker/task
was held in has already been deallocated, it can result in a crash.
Fix the issue by dropping the allocation in the callback instead.

This issue also exists in the stable foundationdb-rs release.