fix integer cast possible undefined behavior
Maybe fdb_iteration should panic or handle the error itself.
`iteration` is checked in foundationdb and returns client_invalid_operation if it is <= 0
There is no information in the API spec about it.
For example if we pass a key.len() < 0 as we could before this commit, it is undefined behavior, because foundationdb is not checking it.