feat(foundationdb,bindingtester): Add directory
Provides structures for managing directories in FoundationDB.

The FoundationDB API provides directories as a tool for managing related
Subspaces. Directories are a recommended approach for administering
applications. Each application should create or open at least one
directory to manage its subspaces. For general guidance on directory
usage, see the discussion in the Developer Guide.

Directories are identified by hierarchical paths analogous to the paths
in a Unix-like file system. A path is represented as a List of strings.
Each directory has an associated subspace used to store its content. The
layer maps each path to a short prefix used for the corresponding
subspace. In effect, directories provide a level of indirection for
access to subspaces.